The init sequence cleanup! Also broke up client.h into a lot of smaller headers. Oh, software rendering works again.
And yes, this does hit a hell of a lot of files.

+++ include/r_dynamic.h
@@ -33,7 +33,7 @@
 
 void R_RocketTrail (vec3_t start, vec3_t end, int type, entity_t *ent);
 void R_RunParticleEffect (vec3_t org, int color, int count);
-void R_RunPuffEffect (vec3_t org, int type, int count);
+void R_RunPuffEffect (vec3_t org, byte type, byte count);
 void R_RunSpikeEffect (vec3_t org, byte type);
 
 void R_BlobExplosion (vec3_t org);
